CDSTRM-1459: Hide NR connect page after signup (#851)
* Add Source parameter for install nerdlet event

This should make it so the Source parameter gets added to the Account
Created event when coming from the install nerdlet.

* CDSTRM-1459: Hide NR connect page after signup

This ensures that after signing up with NR, we don't show the "Connect
to NR" page. Otherwise, that happens because we attempt to follow the
connect link from NR which directs to the connect page if logged in.