CDSTRM-1449: Load all error groups
This changes how we query NR1 for error traces so that we retrieve all
the error groups even when there are thousands of individual error
instances. Instead of querying for all error instances then grouping
them in CS, we now only query for the latest error instance for each
error group, along with the total count of error instances for that
error group.