CDSTRM-1458: Fix slack channel selector issue (#885)
This (hopefully) fixes a reported issue where the channel selector
doesn't work when posting to Slack. The issue seems to be that under
some circumstances we're filtering out the self direct message channel
and thus not populating the user map with the authed user, which results
in an error attempting to access a property of an undefined object. In
theory, this should resolve the error by ensuring that self direct
message channel has an artificially high priority so it never gets
filtered out.